Add StudentIndex to StudentController
