+++ webapp/src/main/webapp/WEB-INF/web.xml
@@ -66,7 +66,7 @@
     <!-- Map all requests for / to the dispatcher servlet -->
     <servlet-mapping>
         <servlet-name>jersey-servlet</servlet-name>
-        <url-pattern>/</url-pattern>
+        <url-pattern>/*</url-pattern>
     </servlet-mapping>
 
     <!-- Error mapping -->
